Cleanup the boolean check logic
the check for a != b || b != c yielded unnecessary extra evaluation
(stuttering to be honest). Instead, first check if a != b. If that is
the case then we can perform the other equality checks based off of
that. If a == b, then we just see if b != c. If that is the case then
just do two more of the six total combos. If b == c then only perform
one execution. The downside of this approach is it causes a small DRY
violation!

This changes seems to provide a several second speed boost at 14 digits!